+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ### HEAD
+* Change roots_sidebar into a more explicit configuration array
 * Re-organize configuration/setup files
 * Update to jQuery 1.8.2
 * Refactor/simplify Roots vCard Widget